usb: Initial commit.
This Pmod is meant to serve as a multipurpose host/client and sniffing
interface, allowing the connection to an FPGA board over a single Pmod
connector of a USB-A and USB-C interface with pass through jumper
capability.

This is based on ideas from @ktemkin and USB Pmod designs by @kbeckmann.